Commented every single property, field, method, class, etc.

+++ Source/InfoShare.Deployment/Data/Actions/File/FileCopyAction.cs
@@ -11,20 +11,34 @@ namespace InfoShare.Deployment.Data.Actions.File
 	/// </summary>
     public class FileCopyAction : BaseAction, IRestorableAction
     {
-		private readonly string _sourcePath;
+        /// <summary>
+        /// The source file path
+        /// </summary>
+        private readonly string _sourcePath;
+
+        /// <summary>
+        /// The destination file path
+        /// </summary>
         private readonly string _destinationPath;
+
+        /// <summary>
+        /// The force switch identifies if file needs to be replaced.
+        /// </summary>
         private readonly bool _force;
 
+        /// <summary>
+        /// The file manager
+        /// </summary>
         private readonly IFileManager _fileManager;
 
-		/// <summary>
-		/// Does copy of file to directory
-		/// </summary>
-		/// <param name="logger"></param>
-		/// <param name="sourcePath"></param>
-		/// <param name="destinationPath"></param>
-		/// <param name="force"></param>
-		public FileCopyAction(ILogger logger, string sourcePath, ISHFilePath destinationPath, bool force = false) 
+        /// <summary>
+        /// Initializes a new instance of the <see cref="FileCopyAction"/> class.
+        /// </summary>
+        /// <param name="logger">The logger.</param>
+        /// <param name="sourcePath">The source file path.</param>
+        /// <param name="destinationPath">The destination file path.</param>
+        /// <param name="force">Replaces existing file if true.</param>
+        public FileCopyAction(ILogger logger, string sourcePath, ISHFilePath destinationPath, bool force = false) 
 			: base(logger)
         {
             _sourcePath = sourcePath;
@@ -35,22 +49,29 @@ namespace InfoShare.Deployment.Data.Actions.File
         }
 
 		/// <summary>
-		/// Executes the action
+		/// Executes the action.
 		/// </summary>
 		public override void Execute()
 		{
 			_fileManager.CopyToDirectory(_sourcePath, _destinationPath, _force);
 		}
 
-		/// <summary>
-		/// Rollback to initial state
-		/// </summary>
-		public virtual void Rollback()
+        /// <summary>
+        /// Reverts an asset to initial state.
+        /// </summary>
+        public virtual void Rollback()
 		{
-			string fileName = Path.GetFileName(_sourcePath);
+			var fileName = Path.GetFileName(_sourcePath);
+
+		    if (string.IsNullOrEmpty(fileName))
+		    {
+		        return;
+		    }
+
 			var copiedFileName = Path.Combine(_destinationPath, fileName);
+
 			if (_fileManager.Exists(copiedFileName))
-		{
+		    {
 				_fileManager.Delete(copiedFileName);
 			}
 		}
